backend/job_log/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
gunicorn>=20.1.0
django-environ>=0.9.0
djangorestframework>=3.13.0
django-cors-headers>=3.13.0
Authlib~=1.0.0
python-dotenv~=0.19.2